+++ admin/menus.php
@@ -59,7 +59,9 @@ if ($action == 'restoreParams')	$result	= oblyon_restore_module ('oblyon');
 // On / Off management
 if (preg_match('/set_(.*)/', $action, $reg)) {
 	$confkey	= $reg[1];
-	$result		= dolibarr_set_const($db, $confkey, GETPOST('value'), 'chaine', 0, 'Oblyon module', $conf->entity);
+	if (preg_match('/^(OBLYON_|THEME_|MAIN_|FIX_|DISABLE_)/', $confkey)) {
+		$result		= dolibarr_set_const($db, $confkey, GETPOST('value', 'alphanohtml'), 'chaine', 0, 'Oblyon module', $conf->entity);
+	}
 }
 // Update buttons management
 if (preg_match('/update_(.*)/', $action, $reg)) {
@@ -105,7 +107,7 @@ print '<script type = "text/javascript">
 						});
 					});
 				</script>
-				<form action = "'.$_SERVER['PHP_SELF'].'" method = "POST">
+				<form action = "'.dol_escape_htmltag($_SERVER['PHP_SELF']).'" method = "POST">
 				<input type="hidden" name="token" value="'.newToken().'" />
 				<input type="hidden" name="action" value="update">
 				<input type="hidden" name="page_y" value="">
@@ -181,6 +183,10 @@ if (getDolGlobalString('MAIN_MENU_INVERT')) {
 						</tr>';
 	}
 }
+// InfraS add begin
+$metas		= array(array(), $conf->entity, 0, 0, 1, 0, 0, 0, '', 'menus');
+oblyon_print_input('OBLYON_TOUCH_MENU', 'on_off', $langs->trans('TouchMenu').'<br><span class = "opacitymedium">'.$langs->trans('TouchMenuHelp').'</span>', '', $metas, 2, 1);	// Touch screen menu mode
+// InfraS add end
 print '				</table>
 				</div>';
 print dol_get_fiche_end();