+++ core/modules/modOblyon.class.php
@@ -54,16 +54,18 @@
 			$this->editor_email				= 'support@inovea-conseil.com';
 			$this->url_last_version 		= 'https://raw.githubusercontent.com/aspangaro/oblyon/14.0/htdocs/custom/oblyon/VERSION';
 			$this->rights_class				= $this->name;																			// Key text used to identify module (for permissions, menus, etc...)
-			$this->family					= 'Inovea Conseil';																		// used to group modules in module setup page
-			$this->module_position			= 10;
-			$this->module_position			= 1;
+			$isDolinfras					= isModEnabled('dolinfras');
+			$family							= $isDolinfras ? getDolGlobalString('DOLINFRAS_FAMILY') : 'Inovea Conseil';
+			$this->family					= $family;																		// used to group modules in module setup page
+			$this->familyinfo				= array($family => array('position' => '001', 'label' => $langs->trans($family)));
+			$this->module_position			= 100016;
 			$this->description				= $langs->trans('Module432573Desc');												// Module description
 			$this->version					= file_get_contents(__DIR__.'/../../VERSION');								// Version : 'development', 'experimental', 'dolibarr' or 'dolibarr_deprecated' or version
 			$this->const_name				= 'MAIN_MODULE_'.strtoupper($this->name);										// llx_const table to save module status enabled/disabled
 			$this->special					= 0;																					// Where to store the module in setup page (0=common,1=interface,2=others,3=very specific)
 			$this->picto					= 'inovea@'.$this->name;																// Name of image file used for this module. If in theme => 'pictovalue' ; if in module => 'pictovalue@module' under name object_pictovalue.png
 			$this->module_parts				= array('menus'	=> 1,
-													'js'	=> array('js'	=> '/'.$this->name.'/js/pushy.js'),
+													'js'	=> array('/'.$this->name.'/js/pushy.js', '/'.$this->name.'/js/oblyon.js'), // InfraS change
 													'css'	=> array('css'	=> ('/'.$this->name.'/css/'.$this->name.'.css'), ('/theme/'.$this->name.'/custom.css.php'), ('/'.$this->name.'/css/font.css')),
 													'tpl'	=> 0,
 													'hooks' => array('data' => array('main'), 'entity' => '0')
@@ -119,21 +121,23 @@
 			$sql		= array();
 			$this->_load_tables('/'.$this->name.'/sql/');
 			oblyon_restore_module($this->name);
+			// InfraS change begin
 			// Copy dir oblyon/themeoblyon to theme/oblyon
-			$srcDir		= dol_buildpath('/oblyon/themeoblyon');
-			$destDir	= DOL_DOCUMENT_ROOT.'/theme/oblyon';
-			if (dol_is_dir($destDir)) {
-				$result	= dol_delete_dir_recursive($destDir);
-				if ($result < 0) {
-					setEventMessage($langs->trans('OblyonDeleteThemeError'), 'errors');
-					return 0;
-				}
-			}
-			$result	= dolCopyDir($srcDir, $destDir, 0, 1);
-			if ($result < 0) {
-				setEventMessage($langs->trans('OblyonCopyThemeError'), 'errors');
-				return 0;
-			}
+			// $srcDir		= dol_buildpath('/oblyon/themeoblyon');
+			// $destDir	= DOL_DOCUMENT_ROOT.'/theme/oblyon';
+			// if (dol_is_dir($destDir)) {
+			// 	$result	= dol_delete_dir_recursive($destDir);
+			// 	if ($result < 0) {
+			// 		setEventMessage($langs->trans('OblyonDeleteThemeError'), 'errors');
+			// 		return 0;
+			// 	}
+			// }
+			// $result	= dolCopyDir($srcDir, $destDir, 0, 1);
+			// if ($result < 0) {
+			// 	setEventMessage($langs->trans('OblyonCopyThemeError'), 'errors');
+			// 	return 0;
+			// }
+			// InfraS change end
 			// Get highest font awesome directory
 			$path				= dol_buildpath('/theme/common/', 0);
 			$listdir			= dol_dir_list($path, 'directories', 0, '^fontawesome-', null, 'name', SORT_ASC, 0, 0, '', 0);
@@ -195,15 +199,16 @@
 
 			dolibarr_del_const($this->db,'MAIN_FONTAWESOME_ICON_STYLE', $conf->entity);
 			dolibarr_del_const($this->db,'MAIN_FONTAWESOME_WEIGHT', $conf->entity);
-
-			$destDir	= DOL_DOCUMENT_ROOT.'/theme/oblyon';
-			if (dol_is_dir($destDir)) {
-				$result = dol_delete_dir_recursive($destDir);
-				if ($result < 0) {
-					setEventMessage($langs->trans('ThemeOblyonErrorDelete'), 'errors');
-					return 0;
-				}
-			}
+			// InfraS change begin
+			// $destDir	= DOL_DOCUMENT_ROOT.'/theme/oblyon';
+			// if (dol_is_dir($destDir)) {
+			// 	$result = dol_delete_dir_recursive($destDir);
+			// 	if ($result < 0) {
+			// 		setEventMessage($langs->trans('ThemeOblyonErrorDelete'), 'errors');
+			// 		return 0;
+			// 	}
+			// }
+			// InfraS change end
 			return $this->_remove($sql, $options);
 		}
 	}